requirements.txt
Django==4.2
djangorestframework==3.14
requests==2.31
newspaper3k==0.2.8
celery==5.3
redis==4.5
gunicorn==20.1
psycopg2-binary==2.9
django-cors-headers==4.2.0